+++ extension/CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to the "al-translation-cleaner" extension will be documented
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
+## [1.0.4] - 2026-04-14
+
+### Changed
+
+* Better Log Outputs for easier use
+* Translations that are commented out arent shown as errors
+* Double-Line Comments are merged to one Line
+
+### Fixed
+
+* Error where the first Enum Value got replaced
+* In PageExt there was Mapping Confusion between modify Action Captions and addAfter Field Captions
+* Double-Line Comments are now recognized
+* 1.0.1 Mapping Error that broke the Extension
+* Names with double Quotes couldnt me Mapped
+
+### Removed
+
+* Option "Ask" in Translation Method, will be added in future updates
+
 ## [1.0.3] - 2026-04-01
 
 ### Fixed